% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedInference.R
\name{enrichedInference}
\alias{enrichedInference}
\title{This function calculates enriched estimates and inference for a high dimensional regression model}
\usage{
enrichedInference(
  X,
  Y,
  k = 1,
  intercept = TRUE,
  standardize = FALSE,
  CVlasso = FALSE,
  Xcorr = c("marginal", "partial"),
  threshold = 0.1,
  sig.level = 0.05,
  Bonf.correction = FALSE
)
}
\arguments{
\item{X}{an nxp matrix of covariates (i.e., n observations and p covariates)}

\item{Y}{a vector of size n containing response values}

\item{k}{number of principal components for the supervised SVD (default is k=1)}

\item{intercept}{intercept for the regression model (default is TRUE)}

\item{standardize}{standardise cavariates (default is FALSE)}

\item{CVlasso}{whether to use CV for lasso tuning parameter selection (default is FALSE)}

\item{Xcorr}{use of either "marginal" or "partial" correlation in the supervised SVD}

\item{threshold}{threshold for the minimum correlation in the supervised SVD (default is 0.1)}

\item{sig.level}{significance level for enriched inferences (default is 0.05)}

\item{Bonf.correction}{Bonferroni correction for enriched multiple testing (default is FALSE)}
}
\value{
enriched estimates of regression parameters for selected covariates and their standard errors,
           enriched confidence intervals, p-values for enriched tests,
           enriched estimate of error variance, estimate of thetak, k, and negative 2log-likelihood of the enriched selected model
}
\description{
This function calculates enriched estimates and inference for a high dimensional regression model
}
